requirements.txt
pandas~=1.5.3
torch~=2.1.2
torchaudio~=2.1.2
omegaconf~=2.2.3
hydra-core~=1.3.2
lightning~=2.2.0